requirements.txt
lightgbm>=4.3.0
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=12.0.0
scikit-learn>=1.4.0
pyyaml>=6.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
joblib>=1.3.0
httpx>=0.28.0
